src/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pydantic[email]==2.10.3
pydantic-settings==2.7.0
aio-pika==9.5.3
pika==1.3.2
redis==5.2.1
boto3==1.35.81
elasticsearch==8.17.0
yfinance>=0.2.54
httpx==0.28.1
lxml==5.3.0
beautifulsoup4==4.12.3
pandas==2.2.3
numpy==2.2.1
apscheduler==3.10.4
python-dotenv==1.0.1
PyJWT==2.9.0
bcrypt==4.2.1
